+++ changelog.md
@@ -4,10 +4,15 @@
 
 * Initial version
 
-### 0.2.0 - 2017-03-19 
+### 0.2.0 - 2017-03-19
 
 * Show sidebar by default
 * Read information about image group parameter and models from web API
 * Select model when running prediction
 * Add more parameter types (e.g., dict, list, enum)
 * Perform simple type checking before submitting model run request
+
+### 0.3.0 - 2017-05-05
+
+* Adopt to changes in the Server API
+* Add visualization widgets for successful model runs